consolidate test functions in testUtils, move asserts to index
