+++ lib/quote.spec.js
@@ -3,32 +3,10 @@ import chaiAsPromised from 'chai-as-promised';
 import proxyquire from 'proxyquire';
 
 import { parseAndGetCrumb } from './yahooCrumb';
-import { assertValidHistoricalResult, assertValidSnapshotResult,
-  getFile, STATIC_CRUMB } from '../tests/testUtils';
-
-// for nodeify to work with stubbed out functions
-import BluebirdPromise from 'bluebird';
+import { stubbedFor } from '../tests/testUtils';
 
 const should = chai.should();
 chai.use(chaiAsPromised);
-
-const yahooCrumbStub = {
-  async getCrumb() { return STATIC_CRUMB }
-};
-
-function utilsDownloadFixture(name) {
-  return {
-    // async download() { return getFile(name); }
-    download: () => BluebirdPromise.resolve(getFile(name))
-  }
-}
-
-function stubbedFor(name) {
-  return {
-    './yahooCrumb': yahooCrumbStub,
-    './utils': utilsDownloadFixture(name)
-  }
-}
 
 describe('quote', () => {
